pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taskbrew"
version = "1.0.6"
description = "Multi-agent AI team orchestrator — coordinate Claude Code, Gemini CLI, and custom AI agents into collaborative development workflows"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Nikhil Chatragadda"},
]
keywords = ["ai", "agents", "orchestrator", "multi-agent", "claude", "gemini", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
    "bandit>=1.7.0",
]

[project.scripts]
taskbrew = "taskbrew.main:cli_main"

[project.urls]
Homepage = "https://github.com/nikhilch98/taskbrew"
Documentation = "https://github.com/nikhilch98/taskbrew/tree/main/docs"
Repository = "https://github.com/nikhilch98/taskbrew"
Issues = "https://github.com/nikhilch98/taskbrew/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
